+++ lib/cartodb/middleware/error/index.js
@@ -1,11 +1,13 @@
 const prepare = require('./prepare');
 const populateTimeouts = require('./populate-timeouts');
-const error = require('./error-middleware');
+const statusCode = require('./status-code');
+const errorResponse = require('./error-response');
 
 module.exports = function errorMiddleware() {
     return [
         prepare(),
         populateTimeouts(),
-        error()
+        statusCode(),
+        errorResponse()
     ];
 };
